+++ fiddles/three/fiddle18Spheres/app.js
@@ -196,8 +196,6 @@
                 cache = app.util.colors.blueGray.cache;
             i = app.util.calcBoundRandomNumber(uBound, cache);
             cache.push(i);
-            console.log('getRandomBlueId');
-            console.log('i = ' + i);
             return i;
         }
     };
@@ -218,15 +216,15 @@
             scene = new THREE.Scene(),
             directionalLight = new THREE.DirectionalLight(0x00EE00, 1.0),
             ambientLight = new THREE.AmbientLight(0x3a66a1),
-            controls = new THREE.TrackballControls(camera),
+        //controls = new THREE.TrackballControls(camera),
             renderLoop = function () {
                 renderer.render(scene, camera);
                 window.requestAnimationFrame(renderLoop);
-                controls.update();
+                //controls.update();
                 app.store.spheres.move();
             };
 
-        renderer.setClearColorHex(0x000000, 1);
+        renderer.setClearColor(0x000000, 1);
         renderer.setSize(WIDTH, HEIGHT);
         app.store.spheres.inflate(19);
         app.store.spheres.render(scene);
@@ -235,7 +233,7 @@
         camera.position.y = -100;
 
         directionalLight.position.set(100, 100, 300);
-        controls.target.set(0, 0, 0);
+        //controls.target.set(0, 0, 0);
         scene.add(directionalLight);
         scene.add(ambientLight);
         scene.add(camera);